Reduce boilerplate code in ruleOnClaim.js
Tests in ruleOnClaim.js contain a large amount of boilerplate at the top
of every test setting up a token and a stake. It will be easier to
refactor this setup code if it's logic is only implemented once, and
easier to add new tests if it is executed automatically for each test.

* Create two new mutable state variables, token and ds
* Move setup boilerplate to a beforeEach
* beforeEach overwrites the token + ds state variables before each test