Add commit-ordered linked-lists to DelphiVoting
In order to implement the new payouts scheme, we need to know which
order arbiters committed votes to their "faction" in, where a faction is
a cohort of arbiters who expressed the same sentiment in some claim. We
can maintain linked lists of arbiters ordered by when they committed for
their faction, and updates to these lists can be done in constant time.

Because we don't know what the votes are until reveal time, this is when
we can start building up the lists. The revealVotes function now takes
an "insert point" argument, which can be derived using the
"getInsertPoint" function. If you're familiar with PLCRVoting, this
works very similarly, except we order by the block number in which the
original commit was made.

The provided insert point is sanity checked (in constant time) in the
EVM, and the EVM reverts if the insert point would un-order the list.

In the final assessment, this commit adds ordered lists of when arbiters
committed for their respective factions. These lists can be referenced
at payout time to deduce owed payout amounts.

* Add the DLL package to ethpm.
* Add a Commit struct to DelphiVoting, which stores the secretHash and
  the block number when the commit was made.
* Add a factions mapping to the Claim struct. This is a mapping of vote
  options to DLLs (each faction has its own DLL).
* Add an insertPoint argument to revealVote
* Add a public view function getInsertPoint
* Add a public view function validPosition, used now by revealVote
* Update tests to use the new revealVote API