Remove O(n) operation in claimFee
This commit replaces a while loop with a recursive function that lazily
computes arbiter ranks and stores them once it has computed them. For a
claim with n arbiters, even if n is very large, all arbiters should be
able to eventually claim so long as they can wait for some arbiters to
claim before them. No single arbiter can block fee claiming for any
other arbiter, and in the best case feeClaiming is O(1).